- Keep track of correct cards - Quiz page for quizing users on their stack

+++ client/app/stack/stack.controller.js
@@ -5,6 +5,7 @@ angular.module('stackrApp')
 
     var stackId = $stateParams.id;
     console.log('stack id', stackId);
+    $scope.card = {};
 
     StacksService.getStack(stackId)
       .success(function (ret) {
@@ -13,16 +14,51 @@ angular.module('stackrApp')
 
     CardsService.getCards(stackId)
       .success(function (ret) {
-          console.log(ret.data);
+          $scope.cards = ret;
       });
 
     $scope.deleteStack = function () {
+      var shouldDelete;
+
+      shouldDelete = confirm("Are you sure you want to delete this stack?");
+      if (!shouldDelete){
+          return;
+      }
+
       StacksService.deleteStack(stackId)
         .success(function (ret) {
             console.log(ret.data);
         })
         .error(function (ret) {
           console.log(ret);
+        });
+    };
+
+    $scope.createCard = function () {
+      var card = {
+        front: $scope.card.front,
+        back: $scope.card.back
+      };
+
+      CardsService.createCard(stackId, card)
+        .success(function (ret) {
+          card.id = ret.id;
+          $scope.cards.push(card);
+        });
+    };
+
+    $scope.deleteCard = function (pCard) {
+
+      var shouldDelete;
+
+      shouldDelete = confirm("Are you sure you want to delete this card?");
+      if (!shouldDelete){
+          return;
+      }
+
+      CardsService.deleteCard(pCard.id)
+        .success(function (ret) {
+          $scope.cards.splice($scope.cards.indexOf(pCard), 1);
         });
 
     };
